refactor: Code improvement by applying SRP (Single Responsibility Principle)
